feat: change error diffusion: weights are duplicated and divided instead of weighed by cell size
The cell size algorithm was more involved to explain and the usage of weights was less intuitive.
Now we just use the same weights configuration but split or merge weights for cells whihc are subdivided resp more or less compared to the checking cell.
A couple of tests reveiled no siginificant impact on the result.